presun cronu notifikace do celery tasku.
